Add ADR for SQLite
Once adl supports it, it'd be great to add the architecture diagram to
the readme, but to do that in a clean way, we'd want to add a subfolder
to the ADR folder, which I don't believe adl supports at this time. We
should circle back and add that diagram and more context when we can.