Cleanup to make room for the cover manager.
Technically, the cover manager is _working_. It's
downloading files, but it's just storing those as cover.jpg
which... isn't the most helpful. I need to think through how
we want to store covers. Maybe this is fine, but we need a route
handler that converts a request to /[isbn]/cover.jpg to load
the actual cover.

I like this idea, and will likely pursue this, but I'll think on it a
bit more.